In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,18 +8,18 @@
 
 int main(void)
 {
-	int first = 1;
-	int second = 2;
-	int next;
-	int count = 0;
+	unsigned long first = 1;
+	unsigned long second = 2;
+	unsigned long next;
+	unsigned long count = 0;
 
-	printf("%d, %d", first, second);
+	printf("%lu, %lu", first, second);
 	count += 2;
 
 	while (count < 98)
 	{
 		next = first + second;
-		printf(", %d", next);
+		printf(", %lu", next);
 		count++;
 
 		first = second;
